+++ packages/linked_layouts/lib/linked_layouts.dart
@@ -3,5 +3,6 @@ library;
 export 'src/frame_callback_scheduler.dart';
 export 'src/layout_client.dart';
 export 'src/layout_link.dart';
+export 'src/layout_links.dart';
 export 'src/render_mixins.dart';
 export 'src/render_objects.dart';